refactor: tsconfig 에서 alias 를 설정 후 반영합니다.
